+++ html/js/music.js
@@ -1,21 +1,15 @@
-import {
-    audioNext,
-    audioPrev,
-    audioWrapper,
-    backgroundAudio,
-} from './util/elements.js';
+import { backgroundAudio } from './util/elements.js';
 import { getHandoverData } from './util/handover.js';
-import { decrementWrap, incrementWrap } from './util/increment-wrap.js';
+import { incrementWrap } from './util/increment-wrap.js';
 import { randomInt } from './util/random.js';
 
 const {
     paths: { music },
-    config: { music: musicEnabled, musicVolume, musicShuffle },
+    config: { music: musicEnabled, initialAudioVolume, musicShuffle },
 } = getHandoverData();
 
 if (musicEnabled && music.length > 0) {
-    backgroundAudio.volume = musicVolume;
-    audioWrapper.style.display = '';
+    backgroundAudio.volume = initialAudioVolume;
 
     /**
      * @param {string} fileName
@@ -29,33 +23,17 @@ if (musicEnabled && music.length > 0) {
         backgroundAudio.loop = true;
         play(/** @type {string} */ (music[0]));
     } else {
-        audioPrev.style.display = '';
-        audioNext.style.display = '';
-
         /** @type {number | null} */
         let currentSong = null;
 
-        const prevSong = () => {
-            currentSong = decrementWrap(currentSong, 0, music.length - 1);
+        const next = () => {
+            currentSong = musicShuffle
+                ? randomInt(0, music.length, currentSong)
+                : incrementWrap(currentSong, 0, music.length - 1);
             play(/** @type {string} */ (music[currentSong]));
         };
 
-        const nextSong = () => {
-            currentSong = incrementWrap(currentSong, 0, music.length - 1);
-            play(/** @type {string} */ (music[currentSong]));
-        };
-
-        const randomSong = () => {
-            currentSong = randomInt(0, music.length, currentSong);
-            play(/** @type {string} */ (music[currentSong]));
-        };
-
-        audioPrev.addEventListener('click', prevSong);
-        audioNext.addEventListener('click', nextSong);
-        backgroundAudio.addEventListener(
-            'ended',
-            musicShuffle ? randomSong : nextSong,
-        );
-        musicShuffle ? randomSong() : nextSong();
+        backgroundAudio.addEventListener('ended', next);
+        next();
     }
 }