+++ html/js/util/increment-wrap.js
@@ -8,14 +8,3 @@ export function incrementWrap(num, min, max) {
         ? min
         : ++num;
 }
-
-/**
- * @param {number | null | undefined} num
- * @param {number} min
- * @param {number} max
- */
-export function decrementWrap(num, min, max) {
-    return typeof num === 'undefined' || num === null || num <= min
-        ? max
-        : --num;
-}